merged, then renamed test.py to what it has become -- the mapper for pokeapi
